scripts/requirements.txt
requests
tqdm
Pillow